Updated directory names, keras lstm updated, directory manager updated, preprocessing updated.
