+++ nn_server/utilities/stacked_bidirectional_lstm.py
@@ -2,35 +2,59 @@ import keras as ks
 import tensorflow as tf
 
 class StackedBidirectionalLSTM:
-    def __init__(self, filter, sampleLength=30, maxWords=100000):
-        self.__sampleLength = sampleLength
-        self.__maxWords = maxWords
+    def __init__(self, filter, vectorizer=None):
         self.__filter = filter
+        self.__vectorizer = vectorizer
+        if vectorizer is None: # if vectorizer not provided, assume filter is vectorizing filter
+            self.__sampleLength = filter.getSampleLength()
+            self.__maxWords = filter.getMaxWords()
+        else: # if vectorizer provided, use the vectorizer
+            self.__sampleLength = vectorizer.getSampleLength()
+            self.__maxWords = vectorizer.getMaxWords()
         self.__model = None
 
-    def initializeNewModel(self):
+    def __initialize(self):
+        print(self.__sampleLength)
+        print(self.__maxWords)
         inputs = ks.Input(shape=(self.__sampleLength,), dtype="int32")
-        layers = ks.layers.Embedding(input_dim=self.__maxWords, output_dim=128, mask_zero=True)(inputs)
+        layers = ks.layers.Embedding(input_dim=self.__maxWords, input_length=self.__sampleLength, output_dim=128, mask_zero=True)(inputs)
         # Stack bidirectional LSTMs
-        layers = ks.layers.Bidirectional(ks.layers.LSTM(512, return_sequences=True))(layers)
-        # layers = tf.keras.layers.Dropout(rate=0.1)(layers)
-        layers = ks.layers.Bidirectional(ks.layers.LSTM(256, return_sequences=True))(layers)
-        # layers = tf.keras.layers.Dropout(rate=0.2)(layers)
-        layers = ks.layers.Bidirectional(ks.layers.LSTM(128, return_sequences=True))(layers)
-        # layers = tf.keras.layers.Dropout(rate=0.3)(layers)
-        layers = ks.layers.Bidirectional(ks.layers.LSTM(64))(layers)
+        layers = ks.layers.Bidirectional(ks.layers.LSTM(units=128, dropout=0.1, return_sequences=True))(layers)
+        layers = ks.layers.Bidirectional(ks.layers.LSTM(units=128, dropout=0.1, return_sequences=True))(layers)
+        layers = ks.layers.Bidirectional(ks.layers.LSTM(units=64, dropout=0.05, return_sequences=True))(layers)
+        layers = ks.layers.Bidirectional(ks.layers.LSTM(units=64))(layers)
         # Add a classifier
         outputs = ks.layers.Dense(units=2, activation="softmax")(layers)
         self.__model = ks.Model(inputs=inputs, outputs=outputs)
-        print("GPU Check: " + str(tf.test.is_gpu_available()))
         self.__model.summary()
+
+    def exportModel(self, filePath):
+        if self.__model is not None:
+            self.__model.load_weights(filePath)
+        else:
+            raise Exception("Cannot export empty uninitialized model.")
+
+    def importModel(self, filePath):
+        self.__initialize()
+        self.__model.save_weights(filePath)
+
+    def importCheckpoint(self, fileDir="./"):
+        self.__initialize()
+        latest = tf.train.latest_checkpoint(fileDir)
+        self.__model.load_weights(latest)
+
+    def trainModel(self, saveFilePath, trainingDataset, validationDataset):
+        self.__initialize()
+        checkpoint = tf.keras.callbacks.ModelCheckpoint(filepath="checkpoint.ckpt", save_weights_only=True, verbose=1)
         self.__model.compile("adam", "binary_crossentropy", metrics=["accuracy"])
-
-    def importModel(self):
-        return 0
-
-    def train(self, dataset):
-        self.__model.fit(dataset, batch_size=32, epochs=10)
+        self.__model.fit(x=trainingDataset[0], y=trainingDataset[1], validation_data=validationDataset, batch_size=256, epochs=5, callbacks=[checkpoint])
+        self.exportModel(saveFilePath)
 
     def process(self, text):
-        preprocessedText = self.__filter(text)
+        if self.__model is None:
+            raise Exception("Cannot use uninitialized model.")
+        prep = self.__filter(text.lower())
+        if self.__vectorizer is not None:
+            prep = self.__vectorizer(prep)
+        pred = self.__model.predict(prep)
+        return pred
